Extract event rendering to event presenter
